move delete to InstallQueueUpdater
This should be the only class that affects the install status.
